Add garden creation (#1)
This change adds a simple golang script which takes all plots in form of
a txt file and places them in the garden, aferwards adds the seed to
fill up space and prints it to its stdout. It can then be used in an
automated fashion to regrow the garden on radicle.xyz.